+++ util/UnitConverter.py
@@ -6,11 +6,12 @@ class ConvertToSystem:
     round_to_decimals = 2
     extract_numbers_pattern = "\d*\.\d+|\d+"
 
-    def __init__(self, system: str):
+    def __init__(self, system: str, full_log=False):
         if system not in self.supported_systems:
             raise ValueError('unit system not supported')
         else:
             self.system = system
+            self.full_log = full_log
 
     def temperature(self, temp_string: str):
         try:
@@ -22,7 +23,8 @@ class ConvertToSystem:
                 return fahrenheit
 
         except Exception as e:
-            print(f'{e}! probably caused by an empty row in the data')
+            if self.full_log:
+                print(f'{e}! probably caused by an empty row in the data')
             return 'NA'
             
     def dew_point(self, dew_point_string: str):
@@ -35,7 +37,8 @@ class ConvertToSystem:
                 return fahrenheit
             
         except Exception as e:
-            print(f'{e}! probably caused by an empty row in the data')
+            if self.full_log:
+                print(f'{e}! probably caused by an empty row in the data')
             return 'NA'
 
     def humidity(self, humidity_string: str):
@@ -44,7 +47,8 @@ class ConvertToSystem:
             return humidity
         
         except Exception as e:
-            print(f'{e}! probably caused by an empty row in the data')
+            if self.full_log:
+                print(f'{e}! probably caused by an empty row in the data')
             return 'NA'
 
     def speed(self, speed_string: str):
@@ -57,7 +61,8 @@ class ConvertToSystem:
                 return mph
 
         except Exception as e:
-            print(f'{e}! probably caused by an empty row in the data')
+            if self.full_log:
+                print(f'{e}! probably caused by an empty row in the data')
             return 'NA'
 
     def pressure(self, pressure_string: str):
@@ -70,7 +75,8 @@ class ConvertToSystem:
                 return inhg
                 
         except Exception as e:
-            print(f'{e}! probably caused by an empty row in the data')
+            if self.full_log:
+                print(f'{e}! probably caused by an empty row in the data')
             return 'NA'
     
     def precipitation(self, precip_string: str):
@@ -83,7 +89,8 @@ class ConvertToSystem:
                 return inches
                 
         except Exception as e:
-            print(f'{e}! probably caused by an empty row in the data')
+            if self.full_log:
+                print(f'{e}! probably caused by an empty row in the data')
             return 'NA'
 
     def uv(self, uv_string: str):
@@ -92,7 +99,8 @@ class ConvertToSystem:
             return measure
             
         except Exception as e:
-            print(f'{e}! probably caused by an empty row in the data')
+            if self.full_log:
+                print(f'{e}! probably caused by an empty row in the data')
             return 'NA'
 
     def solar(self, solar_string: str):
@@ -101,7 +109,8 @@ class ConvertToSystem:
             return measure
             
         except Exception as e:
-            print(f'{e}! probably caused by an empty row in the data')
+            if self.full_log:
+                print(f'{e}! probably caused by an empty row in the data')
             return 'NA'
 
     def clean_and_convert(self, dict_list: list):